dp_lab2-main/requirements.txt
aiohttp==3.9.5
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.28.1
idna==3.7
langchain==0.3.19
langchain-core==0.3.40
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pydantic==2.7.4
qdrant-client==1.8.2
starlette==0.37.2
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
uvicorn==0.29.0
yarl==1.9.4
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
